requirements.txt
Flask==3.0.0
requests==2.31.0
python-whois==0.8.0
dnspython==2.4.2
beautifulsoup4==4.12.2
reportlab==4.0.7
gunicorn==21.2.0
